Work on issue #17
Found the magic number making everything working.
You need a "border" number that seemes to be arbitrary for every block
making it climable, and not glitch in to the block causing you to take
damage. It also appears that it needs to be same for blocks in a chain
you want to climbe from the backside (not same square as block is
placed). In our instance 1/18 seemes to do the trick.